databricks-unity-dxr-integration/pyproject.toml
[project]
name = "databricks-unity-dxr-integration"
version = "0.1.0"
description = "Sync Data X-Ray classifications into Databricks Unity Catalog volumes."
authors = [{ name = "Ohalo Playbooks", email = "support@ohalo.co" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.25.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.0",
    "requests>=2.32.3",
    "tenacity>=8.4.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.12.0",
    "responses>=0.25.0",
    "ruff>=0.5.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
